commit 81c3d08501e57b65e096ec6a2b1d62f87cf8ced7
    12365 pwritev64 can't write at offsets between [2 GiB, 4 GiB)
commit 8d12444d6bc9b8e39d369d7700dcb29b0572abaf
    12353 libproc should also check the GNU build id in addition to the debug link
commit e68846a8e404cfae64e1c4e685626bf6867ce513
    12352 libproc could know about .gnu_debuglink for remote symbol tables
commit a3170057524922242772a15fbeb3e91f5f8d4744
    12350 Add support for Emulex LPe31000 and LPe32000 fibre channel chipsets
    12351 Remove the 32bit DMA restrictions in the emlxs driver

	Conflicts:
	usr/src/uts/common/syscall/rw.c
	usr/src/test/os-tests/tests/Makefile
	usr/src/test/os-tests/runfiles/default.run
	usr/src/lib/libproc/common/Psymtab.c

+++ usr/src/uts/common/syscall/rw.c
@@ -22,11 +22,11 @@
 /*
  * Copyright 2008 Sun Microsystems, Inc.  All rights reserved.
  * Use is subject to license terms.
- * Copyright 2017, Joyent, Inc.
+ * Copyright 2020, Joyent, Inc.
  */
 
 /*	Copyright (c) 1983, 1984, 1985, 1986, 1987, 1988, 1989 AT&T	*/
-/*	  All Rights Reserved  	*/
+/*	  All Rights Reserved	*/
 
 /*
  * Portions of this source code were derived from Berkeley 4.3 BSD
@@ -1087,7 +1087,7 @@ preadv(int fdes, struct iovec *iovp, int iovcnt, off_t offset,
 		}
 	}
 
-	if ((bcount = (ssize_t)count) < 0) {
+	if ((bcount = count) < 0) {
 		if (aiovlen != 0)
 			kmem_free(aiov, aiovlen);
 		return (set_errno(EINVAL));
@@ -1105,7 +1105,7 @@ preadv(int fdes, struct iovec *iovp, int iovcnt, off_t offset,
 	rwflag = 0;
 
 	/*
-	 * Behaviour is same as read(2). Please see comments in read(2).
+	 * Behaviour is same as read(2). Please see comments in read above.
 	 */
 	if (vp->v_type == VREG) {
 		if (bcount == 0)
@@ -1288,7 +1288,7 @@ pwritev(int fdes, struct iovec *iovp, int iovcnt, off_t offset,
 		}
 	}
 
-	if ((bcount = (ssize_t)count) < 0) {
+	if ((bcount = count) < 0) {
 		if (aiovlen != 0)
 			kmem_free(aiov, aiovlen);
 		return (set_errno(EINVAL));
@@ -1306,8 +1306,8 @@ pwritev(int fdes, struct iovec *iovp, int iovcnt, off_t offset,
 	rwflag = 1;
 
 	/*
-	 * The kernel's write(2) code checks the rctl & OFFSET_MAX and returns
-	 * EFBIG when fileoff exceeds either limit. We do the same.
+	 * The kernel's write(2) code checks OFFSET_MAX and the rctl, and
+	 * returns EFBIG when fileoff exceeds either limit. We do the same.
 	 */
 	if (vp->v_type == VREG) {
 		if (bcount == 0)
